ver3, classes seperated
