--- README.md
@@ -1,5 +1,0 @@
-Object List Merger
------
-Coding language: Python3 <br/>
-Coding standard: PEP8 <br/>
-Test framework: unittest <br/>
