--- SOLID.md
@@ -1,2 +1,0 @@
-The main code at this point follows the Open-Closed principle. As we just added a new function called multiSelect to the previous code for covering four inputs. So at this point the code follows the principle and there is no need to implement a specific design pattern.
-About next modifications, the added function is able to handle both two input and four input mode without modifications.
