+++ inputData.py
@@ -1,0 +1,7 @@
+class inputData:
+    def __init__(self,dataNum,objName,objNum,prec):
+        ''' Data type object for storing sensed objects '''
+        self.dataNum = dataNum
+        self.objName = objName
+        self.objNum = objNum
+        self.prec = prec
